Add Example callout and stylized Terminal block for writing user docs
This adds two helpers I was tinkering with while writing user docs.
The first is an "example" callout that stylizes a block of text for
adding examples of how to do something.

The second is a stylized Terminal block that's a bit more friendly than
trying to triple backtick everything. I was especially annoyed by this
because I like angle brackets <> for denoting when users would have to
provide their own input, and those are colored incorrectly when you do
triple backticks with bash formatting.